Add func and test to get n-oct spaced frequencies. Can't work.
Can't use the mathematical expression to get the frequencies, because
the real thir-octave spacing, for example, is not exactly the same as
the one that is commonly used. What should be done, actually, is to hard
code values and then find nearests ones to the ones that are asked.

Octave spaces, half, and third would be sufficient.