Add spectral subtraction function with test.
It's a correct implementation of the spectral subtraction function, like
in the sEPSM, but think it could be "improved". For example, it does not
use the period window for the hanning window.

Also, I do not use the same indexes to set the noise estimate to zero;
it does make sense to me. Why would I set to zero the same indexes as in
the speech signal, if the whole point is to keep values above zero. It
creates errors when doing the square root. However, it does make sense
to apply exactly the same processing to the noise and to the speech.