Fix third octave filtering to work directly on spectrum
This is a direct port of the sEPSM version1 code, that Soren gave me on
2013-10-11 or so. The test passes, but the relatice tolerance is set to
1%, which is quite high. Somehow the results are not exactly the same,
even if the signals are the same and the frequency indeces are the same
as well. Should double check.
Also, the code now uses a FFT of the size of the signal, instead of a
power of 2. I should probably fix that, with the changes involved in the
normalization, because is was a pretty severe bottleneck before.