Select bands above threshold using the mixture.
From an email from Soren on 2013-10-17:

Yes, in an earlier version I used the noise alone to check against the
threshold, because I found that this "path" of the model led to errors,
which did not show op in the mixture-path. I found the need to
"objectively" remove the bands that produced the errors and introduced
the threshold check.
I actually always thought that it was kind of counter-intuitive that the
threshold check was performed on the noise alone, and saw it more as a
fix in the model.
However, I found that the errors were specific to some conditions with
spectral subtraction. At the time i was estimating the processed noise
alone via a kind of "shadow filtering" where I effectively applied the
same gain function to the mixture and the noise alone. But to be
consistent with the methodology used in other investigation of mine,
including simulations with ESII, I changed this to use the Hagerman
separation method. The noise-alone estimated via the two approaches are
not exactly identical, but effectively led to the same model
predictions. But the funny thing was that now I could do the threshold
check on the mixture, which I think is more logical. I don't think the
real system makes a separate threshold "check" for the speech and noise
parts. Moreover, performing the check on the noise alone created
problems when simulation with speech distorted by phase jitter and an
infinite SNR, where there was no external noise.