Fix level setting and spectral subtraction processing
I just used the level setting that was in the Matlab sEPSM, even if I
don't think it's particularly elegant. But at least, now both
implementation give the same results.

The spectral subtraction problem was that the input parameter to the
OverAdd function is the square root of some negative values. In Matlab,
the sqrt of a negative value gives a complex number. In Python, if the
input is real, then the output if NaN, but if the input is complex, then
the output is also complex. So I simply forced that input to the square
root to be complex.