Remove gain of 2 when doing gammatone filtering.
Remove the factor 2 that multiplicated the output of the gammatone
filtering. I never understood why it was there in the first place. It
means that the peak of the gammatone filter has a gain of 2 rather than
1. So far, the sEPSM makes the same prediction, because the SNRenv is a
ratio, but it means that the scale of everything is half of what it was
before.