Find toolbox version without importing it.
Changed the code to find the toolbox version so that it uses a function
to find the version number in pambox/__init__.py rather than importing
pambox.

It corrects a bug where it would not be possible to install pambox from
setup.py install if the requirements were not already installed. I read
that it was a problem on JeffKnupp.com; he was making the same mistake
in his package `sandman`.