Replared OrderedDict with array.
I was using OrderedDicts in the mr-sEPSM as a way of saving the
filtered envelopes, which are of different lengths. I replaced them with
a simple list. I'm not sure it'll make anything clearer in the long run,
but it seemed like a better data structure when combined with numpy data
structures.