Make length of downsampled envelope parametric.
Made the length of the initialized downsampled envelope a parametric
depending on if downsampled envelope is odd of even. I know that the
mod_filterbank function will return an odd-length vector, so I have to
plan to output accordingly. I know it's not super elegant, but I'm
keeping it that way until I optimized the mod_filterbank function to use
padded FFT anyway.