Change mr-sEPSM to use masked arrays
I replace the data structure to store the multi-resolution modulation
envelope powers to be masked arrays, rather than list of arrays. This
way, we can directly apply np.minimum, np.mean, or other functions
without having to loop through a given dimensions.

However, I had to change the code to use data created from python,
because the Matlab data could not be used, since it's not masked array.

The code is about 20 times faster for the calculation of the modulation
power spectrum.