Extract modulation filtering from SNRenv.
The modulation filterbank was included in the snr-env function because
that's how it was in the Matlab implementation. It didn't make sense,
however, because the point of the SNRenv process is to do the ratio of
(S-N)/N.
I had to change the test. I wrote them by hand. They could clearly be
more elaborate.