Make variable naming more consistent in mr-sEPSM
The variable names were all over the place in the mr-sEPSM. I changed
them to be more consistent among themselves, since there are multiple
representation that are almost equivalent, like the multi-resolution
SNRenv, and the time-averaged multi-resolution SNRenv.

It changed the output slighlyt, because there are now more variable
names in the output named tuple. I should really do a cleanup of this
and use dictionnaries for each type of information.
#
# Why is this change necessary?
# How does it address the issue?
# What side effects does this change have?
# Consider including a link \[to the discussion.\]