Use rfft and output filtered signals.
In noctave_filterbank, output the time domain signal as well as the rms
values. Arguably, we should only output the time signals, and apply the
RMS calculations as a second step...

Also, replace the calls to FFT with calls to RFFT because we know that
signals are going to be speech.

Had to make a simple change to the tests because there are now two
outputs to the function.