Working mr-sEPSM with sliding mr-window.
Implementation of the mr-sEPSM with a sliding window for the computation
of the multi-resolution envelope power calculation. The step size is set
by the smallest window, associated with the highest modulation frequency
(e.g. 1/256 ms). The window is centered at the index of the step. This
means that the first windows are incomplete for each modulation center
frequency, because half of the window falls "out" of the array because
it is "before" the start of the signal. This is not so noticeble for
high modulation frequencies, but it is easier to observer for the lowest
modulation frequency, which could have only 1 or 2 full windows for a
given sentence.

The current implementation is rather slow. The sliding window is
implemented as a generator. I can kind of thing of a way of doing it
with numpy's strides or reshape, but I won't implement it now.