Add fs parameter to Westermann condition
Added a parameter to set the sampling frequency of the material for the
Westermann condition. By default it uses a sampling frequency of 40000
Hz, since it is what the material was originally recorded for. I made a
down sampled version of it, at 22050 Hz, to use with other speech
materials at that sampling frequency.