Models return a dictionary instead of a namedtuple
Namedtuple were incompatible with pickling when using IPython.parallel.
The original idea of the namedtuple was that the return data should not
be modified, but I don't think the cost of using a mutable type is that
high, considereing that the gain is now that the models are compatible
with IPython.parallel.