Allow setting sigma_s when fitting ideal observer.
The sigma_s parameter is usually fixed for a given language, but there
was no was previously to force the sigma parameter to be fixed during
the optimization.

The implementation is rather simple; i just defined a new cost function.
It was have been much nicer to define a cost function what would take an
arbitrary selection of parameters to fix, but I could not figure out how
to do that, since the optimization function only provides a vector to
the optimization function, without defining which parameter is which.
There is probably a way to do what I want do with a cost function
defined as a closure, or something like that.