Add option to pad first signal or not
Added an option to select whether the fist signal should be zero padded,
or if the second one should be truncated. This is just a convenience
function, for when two signals should be the same length, but you don't
want to zero-pad the first one because doing so would change its
envelope, and therefore its modulation spectrum.