Add options to plotting functions
Added a bunch of options, mostly to set minimum and maximum values of
the colormap, and the set the axes or figures in which to plot the
figure. The changes were made so that it's easier to make consistent
looking figures and to allow the composition of the figure elements into
larger figures.