Define Hilbert transform done on last dimension.
I tried implementing the Hilbert envelope such that the envelope could
be calculated on any dimensions of the array. I ran into some problems
while doing that, mainly on how to make sure the array returned is the
same dimensions as the original. For now, it just defined it as though
the envelope extraction had to be done on the last dimension, which is
reasonable for now, and consistent with the rest of the input/output
handling in the rest of the toolbox.