MrSepsm: Select mod channels lower than 1/4 chan.
When doing the optimal combination of SNRenv values, in the
time-averaged SNRenv matrix, only select the SNRenv for modulation
frequencies that are lower than 1/4 of the cetner frequency of the
peripheral filter. For example, for the 100 Hz filter, only select the
modulation channels up to, and including, 16 Hz.

This step is included in the publised mr-sEPSM, as well as in the code,
but I had missed it.