Use singla codebase for Python 2 and 3
Uses the `six` package for some of the imports.

Also, I moved to using mat files to store test data.
I was previously using the output of np.dump to store MaskedArrays, but
that was causing a lot of trouble when trying to read it in Python 2 or
3. I saved the data to Matlab mat files instead. Each file has a 'data'
and 'mask' value that must be combined using

    np.ma.MaskedArray(mat['data'], mat['mask'])

It's a little of extra typing, but now the format is version, and
platform independent, and pambox doesn't need to depend on an extra
package, like PyTables of H5DF to save arrays.

Fix imports for having a single codebase