Make optinal time-output of noctave filterbank
In `noctave_filterbank, creating the time output of the filterbank was
done by default. It turns out that the `irfft` function from numpy is
not linked to MKL, unlike `fft`, `rfft`, and `ifft`, such that it is
actually quite expensive to calculate the inverse FFT on large arrays.

Right now, I just made the time output optional. It's actually a bit
strange to make optional the temporal output of a filterbank... It would
probably be a better idea to simply replace the call to `irfft` with
`ifft`, and to create the correct vector myself...