Increase tolerance for ideal obs test
Somehow, although the tests pass on my machine, they do not pass on
Travis-ci. The values that the ideal-obs fit are slightly different.
It's probably something with the version of the libraries on Linux and
OS X? I chose to increase the tolerance slightly, from a relative
tolerance of 1e-3 to an absolute tolerance of 1e-4. Changes smaller than
this are unlikely to create any different in the actual values
transformed by the ideal observer.