Refactor toolbox
Split toolbox in parts of the auditory system. There an "inner",
"middle" and "outer" part, as well as a "central". The idea is to mirror
the auditory system. The inner, middle, and outer parts are not put into
a "peripheral" module because that would be an unnecessary level of
nesting.

Besides those two modules, there are also a utils module and a speech
module. There could also be "monoral" or "binaural" modules, although,
arguably, they could also be in the "central" module.