Fix #9: Standardized intelligibility models' prediction output
I settle to having two levels of nested dictionaries at the output of
the intelligibility models.
It is a bit verbose to access the results, but the upside is that it is
straight forward to store multiple outputs in a dataframe.

It is possible to simply loop over the ['p'] key of the results.