Create output directory if it doesn't exist
Checks if the output directory exists before writing the output data.
I'm not sure how good an idea this is, but it seems safe to me to check
first. Otherwise, the default is to try to write the data to the current
directory. It is not a bad idea, but it can rapidly create a mess.