Add optional models parameters to pred_to_pc
I added an optional parameter to the speech/experiment.pred_to_pc
function to specify the model to transform. Without this parameter, all
the rows were transformed using the same function, which is not really
useful.

The inputs can be a single model name, a list of model names, or a
dictionary with model names and output names. It think it offers the
most flexibility.

Another option would be to accept models directly, and then check their
`name` attribute. That would make things more straight forward, if the
users already have the models created, but it would be a bit awkward if
they are doing transforms from text file they just loaded...
I don't think it would hurt to add it...