Fix #14 which made mod_filterbank acausal
Removed the `-1` factor when creating the frequency vector for
filtering. The results is that the filter is causal and produces the
same time output as using the scipy.signal.filtfilt with butterworth
filter coefficients.

Actually, replacing the implementation with filtfilt might be a good
idea, since the shape of the filters would not be dependent on the
length of the signal.