Add method to extract groups
The method returns the groups over which the experiment is run, like the
models, the distortion parameters, SNR, etc., so that we can do some
processing (averages) across conditions. It ignores columns like values
and intelligibility.

It's probably too complex and could be simplified. It could also be
taken out of the Experiment class. The way it is now, it's expected that
we know the experiment class when we analysis results.