Fix #17: Remove bandwidth compensation constant
When finding the bands above threshold in the sEPSM, there is a factor
of 0.231 for the compensation of the filter bandwidth. This factor is
unnecessary because the diffuse hearing threshold used for the
comparison are already adjusted for the filter bandwidths.

The tests tolerance had to be adjusted; for the spectral subtraction case,
the relative difference compared to the Matlab code is smaller than 8%. In
the condition with speech-shaped noise only, the difference is smaller than 0
.1%.