Add function to calculate SRTs from predictions
It works if the SRTs calculation function applies to all groups (e.g.
Model, Output, Condition), which is usually the case. However, the
function does not work if the parameters are different for each
condition, such as for the STEC model (Wan et al. 2010 and 2014). In
that case, it would be possible to loop over the different parameters
values required for the conversation to SRT and to call this function
repeatedly.