Fix bug when selecting valid modulation channels
The mr-sEPSM paper states that only modulation filters that have center
frequencies below one fourth of the peripheral filter's center frequency
are considered. The "1/4" parameters was missing from the code.

It has a small impact on the mr-sEPSM prediction. I had to change the
target value for the test by 0.25.