Change path of all engines to current path
This makes sure that all relative paths in the experiment keep on
working. It fixed a few problems, for example if there's a module in the
current path, but that is not part of the python path. Essentially, is
to account for the case where the engines were not started in the folder
where the experiment is ran.