Add optional parameter to define output path
The idea is that it's possible to define ahead of time where the results
of an experiment will be saved. This way, it's possible to refer to the
path later on in a script/application.

It could have been another return value to the `run` function, however,
since `run` doesn't always save to disk, then the path could be empty or
None. It would also make the number output value change.