API: Make models' predict inputs kwargs
Change the arguments to the intelligibility models to keyword arguments.
It allows models that require only one of the signals (say, mixture) to
be called as such:

    model.predict(mix=noisy)

instead of requiring dummy values, such as:

    model.predict(None, noisy, None)

The change is backwards compatible with previous code, because the order
of the arguments did not change.