Include masker in condition list
Should solve the issue that the masker changed with every
model/snr/parameters set, because it was loaded *after* the product
across all conditions. Now it's loaded at first, with the target, such
that they're paired for the duration of the experiment. It makes the
experiment more different than what is done with humans, but at the same
time, I would argument humans are not as sensitive to target-masker
pairing variations.