ENH: Add B-sEPSM from Chabot-Leclerc et al. (2016)
This adds the B-sEPSm model, which is a binaural extension to the
mr-sEPSM model of Jorgensen et al. (2013). The model was proposed by
Chabot-Leclerc et al. (2016, JASA).

The model relies on the mr-sEPSM and on the Equalization--Cancellation
process proposed by Wan et al. (2014, JASA). An implementation of that
EC process is also added here.